backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
exceptiongroup==1.2.2
fastapi==0.115.4
flashtext==2.7
idna==3.10
pydantic==2.9.2
pydantic_core==2.23.4
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
pymysql==1.1.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
urllib3==2.2.3
SQLAlchemy==2.0.36
